+++ src/cargo.rs
@@ -3,18 +3,31 @@ use reqwest::header::USER_AGENT;
 use serde::Deserialize;
 
 use crate::config::Config;
+use crate::VERSION;
 
 // CRATES.IO URLS
 const VERSIONS_URL: &str = "/v1/crates/tod/versions";
 
 #[derive(Deserialize)]
 struct CargoResponse {
-    versions: Vec<Version>,
+    versions: Vec<CargoVersion>,
 }
 
 #[derive(Deserialize)]
-struct Version {
+struct CargoVersion {
     num: String,
+}
+
+pub enum Version {
+    Latest,
+    Dated(String),
+}
+pub fn compare_versions(config: Config) -> Result<Version, String> {
+    match get_latest_version(config) {
+        Ok(version) if version.as_str() != VERSION => Ok(Version::Dated(version)),
+        Ok(_) => Ok(Version::Latest),
+        Err(err) => Err(err),
+    }
 }
 /// Get latest version number from Cargo.io
 pub fn get_latest_version(config: Config) -> Result<String, String> {